pyproject.toml
[tool.poetry]
name = "testing-speech-llm"
version = "0.1.0"
description = ""
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^5.15.0"
numpy = "^2.2.2"
soundfile = "^0.13.1"
google-generativeai = "^0.8.4"
pydantic = "^2.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
